requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pypdf
pypdf2
chromadb
faiss-cpu
langchain_google_genai
plangchain-community
openpyxl
unstructured